requirements.txt
python-telegram-bot==20.7
transformers==4.36.2
torch==2.1.2
peft==0.7.1
accelerate==0.25.0
bitsandbytes==0.41.3
sentencepiece==0.1.99
protobuf==4.25.1
faiss-cpu==1.7.4
sentence-transformers==2.2.2
requests==2.31.0
aiohttp==3.9.1
PyPDF2==3.0.1
python-docx==1.1.0
pillow==10.1.0
numpy==1.26.2
datasets==2.15.0
huggingface-hub==0.20.1
